+++ bin/create-user.js
@@ -2,11 +2,12 @@
 
 /**
  * to run create-user applet from the project root folder type in your console:
- * > node bin/create-user <client_id> <client_secret> <email> <password> <first_name> <last_name>
+ * > node bin/create-user <client_id> <client_secret> <email> <password> <first_name> <last_name> <number>
  * <client_id>, <client_secret>, <email>, <password> - are required params
- * <first_name>, <last_name> - are optional
+ * <first_name>, <last_name>, <number> - are optional
  * options:
  * --verify-email - send verification email
+ * --skip-trial - skip 30 day free trial
  */
 
 'use strict';
@@ -22,9 +23,11 @@ const [
   password,
   first_name,
   last_name,
+  number,
 ] = params;
 
 const verifyEmail = flags.includes('--verify-email');
+const skip_30day_trial = flags.includes('--skip-trial') ? 1 : 0;
 
 const { promisify } = require('../utils');
 const { user: { create: createUser } } = require('../lib')({
@@ -39,6 +42,8 @@ createUser$({
   password,
   first_name,
   last_name,
+  number,
+  skip_30day_trial,
   options: { verifyEmail },
 })
   .then(res => console.log(res))