+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Added optional ability to send verification email in *Create User* feature
+- Added `number` and `skip_30day_trial` parameters to *Create User* feature
 
 ## [v1.6.0] - Unreleased Phase 5